// src/app/real-time/models/settings.model.ts
export class Settings {

    uniqueId: string;
    userUniqueId: string;  
    mailOptin: string;  
    mailUpdatedAt: string;
    smsOptin: string;  
    smsUpdatedAt: string;  
    whatsappOptin: string;  
    whatsappUpdatedAt: string;  
    notificationsOptin: string;  
    notificationsUpdatedAt: string;  
    locationOptin: string;  
    locationUpdatedAt: string;  
    payload: string;
    status: string;  
    enabled: string;

}